+++ R/get.measure_2d.R
@@ -105,8 +105,7 @@ get.measure_2d<- function(fitnessMeasure,measures,out.lm, data, dists,aR,radii,r
     if (isS4(out.lm)) {
       stop('Fitness measure not supported for multinomial.  Please use AIC, AICc or BIC')
     } else {
-      set.seed(cv.opts$cv.gamMRSea.seed)
-      fitStat <- cv.gamMRSea(data, out.lm, K=cv.opts$K, cost=cv.opts$cost)$delta[2]
+     fitStat <- cv.gamMRSea(data, out.lm, K=cv.opts$K, cost=cv.opts$cost, s.eed = cv.opts$cv.gamMRSea.seed)$delta[2]
     }
   }