+++ R/getCVfoldID.R
@@ -5,9 +5,9 @@
 #' @param data data used in regression model
 #' @param folds integer number of validation data sets
 #' @param block column in data indicating the blocking structure for cross-validation (if \code{block} = NULL, individual observations will be used as blocks)
-#' @param seed integer number used to set the seed of the fold generation.  By default this is set to `1234`.
+#' @param seed integer number used to set the seed of the fold generation.  By default this is set to `NULL`.
 #'
-#' @details The function returns a random sequence of 1:folds of the same length as observations in data.
+#' @details The function returns a random sequence of 1:folds of the same length as observations in data. The seed used for generation is stored in the attributes (`s.eed`). 
 #'
 #' @examples
 #' # load data
@@ -20,7 +20,12 @@
 #' @export
 #' 
 #' 
-getCVids <- function(data, folds, block=NULL, seed=1234){                        
+getCVids <- function(data, folds, block=NULL, seed=NULL){   
+  
+  if(is.null(seed)){
+    seed<-sample(1:100000, size = 1)
+  }
+ 
   if(is.null(block)==T)
   {
     N <- 1:nrow(data)                                       
@@ -37,6 +42,11 @@ getCVids <- function(data, folds, block=NULL, seed=1234){
       blocks<-unique(block)
     }
     nBlocks<-length(blocks)
+    
+    if(nBlocks < folds){
+      stop("Not enough unique blocks to make K folds. Please reduce folds and try again")
+    }
+    
     n_cv <- ceiling(nBlocks/folds)
     set.seed(seed)                                          
     id_block_cv <- sample(rep(1:folds,n_cv),n_cv*folds)
@@ -47,5 +57,6 @@ getCVids <- function(data, folds, block=NULL, seed=1234){
       id_cv[rows]<-id_block_cv[xi] 
     }
   }
+  attr(id_cv, "s.eed") <- seed
   return(id_cv)                                           
 }